Makefile to support open source tools.
This adds support for arm-none-eabi- tools on Linux instead of the ST
tools. Only one change was required to the sources for supplying C
standard integer types.

Builds to flash.bin, which can be programmed using the open source
st-link tools and loaded on the DERULER and works as expected.